+++ apps/toolbox/src/app-root.xml
@@ -1,2 +1,2 @@
-<Frame defaultPage="pages/transitions">
+<Frame defaultPage="main-page">
 </Frame>
