+++ packages/core/ui/index.ts
@@ -80,5 +80,6 @@ export { PageTransition } from './transition/page-transition';
 export { FadeTransition } from './transition/fade-transition';
 export { SlideTransition } from './transition/slide-transition';
 export { SharedTransition, SharedTransitionAnimationType } from './transition/shared-transition';
+export type { SharedTransitionConfig } from './transition/shared-transition';
 export { WebView } from './web-view';
 export type { LoadEventData, WebViewNavigationType } from './web-view';
