+++ packages/core/ui/transition/shared-transition-helper.android.ts
@@ -1,0 +1,7 @@
+import { SharedTransitionState } from './shared-transition';
+
+export class SharedTransitionHelper {
+	static animate(state: SharedTransitionState, transitionContext: any, type: 'page' | 'modal') {
+		// may be able to consolidate android handling here in future
+	}
+}
